# specs/002-detail-and-viral/contracts/api-spec.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Industry Intelligence API",
        "version": "1.1.0"
    },
    "paths": {
        "/api/feed": {
            "get": {
                "summary": "Get Intelligence Feed (List)",
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/IntelligenceCard"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/article/{id}": {
            "get": {
                "summary": "Get Article Detail",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ArticleDetail"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Polarity": {
                "type": "string",
                "enum": [
                    "positive",
                    "negative",
                    "neutral"
                ]
            },
            "Impact": {
                "type": "object",
                "properties": {
                    "entity": {
                        "type": "string"
                    },
                    "trend": {
                        "type": "string",
                        "enum": [
                            "up",
                            "down"
                        ]
                    },
                    "reason": {
                        "type": "string"
                    }
                },
                "required": [
                    "entity",
                    "trend",
                    "reason"
                ]
            },
            "IntelligenceCard": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "polarity": {
                        "$ref": "#/components/schemas/Polarity"
                    },
                    "fact": {
                        "type": "string"
                    },
                    "impacts": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Impact"
                        }
                    },
                    "opinion": {
                        "type": "string"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "source_name": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "title",
                    "polarity",
                    "fact"
                ]
            },
            "ArticleDetail": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/IntelligenceCard"
                    },
                    {
                        "type": "object",
                        "properties": {
                            "content": {
                                "type": "string",
                                "description": "Markdown or HTML body"
                            },
                            "summary": {
                                "type": "string",
                                "description": "The Brain AI Summary"
                            },
                            "original_url": {
                                "type": "string"
                            }
                        },
                        "required": [
                            "content"
                        ]
                    }
                ]
            }
        }
    }
}